Separte all the part of the decorator
