+++ capatalized_decorator.rb
@@ -1,0 +1,7 @@
+require './base_decorator'
+
+class CapitalizeDecorator < BaseDecorator
+  def correct_name
+    @name_able.correct_name.capitalize
+  end
+end
