+++ name_able.rb
@@ -1,0 +1,5 @@
+class Nameable
+  def correct_name
+    raise NotImplementedError, 'The correct name has not been implemented yet'
+  end
+end
